+++ modules/core/src/test/java/org/locationtech/jts/io/twkb/TWKBTestSupport.java
@@ -9,7 +9,7 @@
  *
  * http://www.eclipse.org/org/documents/edl-v10.php.
  */
-package org.locationtech.jts.io;
+package org.locationtech.jts.io.twkb;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
@@ -24,6 +24,9 @@ import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVRecord;
 import org.junit.rules.ExternalResource;
 import org.locationtech.jts.geom.Geometry;
+import org.locationtech.jts.io.ParseException;
+import org.locationtech.jts.io.WKBReader;
+import org.locationtech.jts.io.WKTReader;
 
 import lombok.Builder;
 import lombok.Cleanup;
